[refactor] MainPage 컴포넌트 분리, share 수정
